FIR/UAST: commonize parameter creation in UMethod
