+++ uast/uast-kotlin-fir/src/org/jetbrains/uast/kotlin/declarations/FirKotlinSecondaryConstructorWithInitializersUMethod.kt
@@ -3,24 +3,20 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.uast.kotlin.declarations
+package org.jetbrains.uast.kotlin
 
 import org.jetbrains.kotlin.asJava.elements.KtLightMethod
 import org.jetbrains.kotlin.psi.KtCallElement
 import org.jetbrains.kotlin.psi.KtClassOrObject
 import org.jetbrains.uast.UElement
 import org.jetbrains.uast.UExpression
-import org.jetbrains.uast.UParameter
 import org.jetbrains.uast.UastEmptyExpression
-import org.jetbrains.uast.kotlin.BaseKotlinSecondaryConstructorWithInitializersUMethod
-import org.jetbrains.uast.kotlin.lz
 
 class FirKotlinSecondaryConstructorWithInitializersUMethod(
     ktClass: KtClassOrObject?,
     psi: KtLightMethod,
     givenParent: UElement?
-) : BaseKotlinSecondaryConstructorWithInitializersUMethod(ktClass, psi, givenParent), FirKotlinUMethodParametersProducer {
-    override val uastParameters: List<UParameter> by lz { produceUastParameters(this, receiverTypeReference) }
+) : BaseKotlinSecondaryConstructorWithInitializersUMethod(ktClass, psi, givenParent) {
 
     override fun buildDelegationCall(delegationCall: KtCallElement, uastParent: UElement): UExpression {
         return UastEmptyExpression(uastParent)
