+++ uast/uast-kotlin-fir/src/org/jetbrains/uast/kotlin/declarations/FirKotlinUAnnotation.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.uast.kotlin.declarations
+package org.jetbrains.uast.kotlin
 
 import com.intellij.psi.PsiClass
 import com.intellij.psi.PsiElement
@@ -11,7 +11,6 @@ import com.intellij.psi.ResolveResult
 import org.jetbrains.kotlin.asJava.toLightAnnotation
 import org.jetbrains.kotlin.psi.KtAnnotationEntry
 import org.jetbrains.uast.*
-import org.jetbrains.uast.kotlin.KotlinAbstractUElement
 
 class FirKotlinUAnnotation(
     annotationEntry: KtAnnotationEntry,
