+++ uast/uast-kotlin-fir/src/org/jetbrains/uast/kotlin/declarations/FirKotlinUMethod.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.uast.kotlin.declarations
+package org.jetbrains.uast.kotlin
 
 import com.intellij.psi.PsiMethod
 import org.jetbrains.kotlin.asJava.LightClassUtil
@@ -12,20 +12,16 @@ import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.psi.psiUtil.containingClassOrObject
 import org.jetbrains.kotlin.psi.psiUtil.getParentOfType
 import org.jetbrains.uast.*
-import org.jetbrains.uast.kotlin.BaseKotlinUMethod
-import org.jetbrains.uast.kotlin.getKotlinMemberOrigin
 
 class FirKotlinUMethod(
     psi: PsiMethod,
     sourcePsi: KtDeclaration?,
     givenParent: UElement?
-) : BaseKotlinUMethod(psi, sourcePsi, givenParent), FirKotlinUMethodParametersProducer {
+) : BaseKotlinUMethod(psi, sourcePsi, givenParent) {
     constructor(
         psi: KtLightMethod,
         givenParent: UElement?
     ) : this(psi, getKotlinMemberOrigin(psi), givenParent)
-
-    override val uastParameters: List<UParameter> by org.jetbrains.uast.kotlin.lz { produceUastParameters(this, receiverTypeReference) }
 
     companion object {
         fun create(
@@ -37,7 +33,7 @@ class FirKotlinUMethod(
                 kotlinOrigin is KtConstructor<*> ->
                     FirKotlinConstructorUMethod(kotlinOrigin.containingClassOrObject, psi, givenParent)
                 kotlinOrigin is KtParameter && kotlinOrigin.getParentOfType<KtClass>(true)?.isAnnotation() == true ->
-                    FirKotlinUAnnotationMethod(psi, givenParent)
+                    KotlinUAnnotationMethod(psi, givenParent)
                 else ->
                     FirKotlinUMethod(psi, givenParent)
             }
