+++ uast/uast-kotlin/src/org/jetbrains/uast/kotlin/declarations/KotlinSecondaryConstructorWithInitializersUMethod.kt
@@ -15,9 +15,7 @@ class KotlinSecondaryConstructorWithInitializersUMethod(
     ktClass: KtClassOrObject?,
     psi: KtLightMethod,
     givenParent: UElement?
-) : BaseKotlinSecondaryConstructorWithInitializersUMethod(ktClass, psi, givenParent), KotlinUMethodParametersProducer {
-
-    override val uastParameters by lz { produceUastParameters(this, receiverTypeReference) }
+) : BaseKotlinSecondaryConstructorWithInitializersUMethod(ktClass, psi, givenParent) {
 
     override fun buildDelegationCall(delegationCall: KtCallElement, uastParent: UElement): UExpression {
         return KotlinUFunctionCallExpression(delegationCall, uastParent)
