--- uast/uast-kotlin/src/org/jetbrains/uast/kotlin/declarations/KotlinUAnnotationMethod.kt
@@ -1,17 +1,0 @@
-/*
- * Copyright 2010-2021 JetBrains s.r.o. and Kotlin Programming Language contributors.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
- */
-
-package org.jetbrains.uast.kotlin
-
-import org.jetbrains.kotlin.asJava.elements.KtLightMethod
-import org.jetbrains.uast.UElement
-
-class KotlinUAnnotationMethod(
-    psi: KtLightMethod,
-    givenParent: UElement?
-) : BaseKotlinUAnnotationMethod(psi, givenParent), KotlinUMethodParametersProducer {
-
-    override val uastParameters by lz { produceUastParameters(this, receiverTypeReference) }
-}
