+++ uast/uast-kotlin/src/org/jetbrains/uast/kotlin/declarations/KotlinUMethod.kt
@@ -28,13 +28,11 @@ open class KotlinUMethod(
     psi: PsiMethod,
     sourcePsi: KtDeclaration?,
     givenParent: UElement?
-) : BaseKotlinUMethod(psi, sourcePsi, givenParent), KotlinUMethodParametersProducer {
+) : BaseKotlinUMethod(psi, sourcePsi, givenParent) {
     constructor(
         psi: KtLightMethod,
         givenParent: UElement?
     ) : this(psi, getKotlinMemberOrigin(psi), givenParent)
-
-    override val uastParameters by lz { produceUastParameters(this, receiverTypeReference) }
 
     companion object {
         fun create(psi: KtLightMethod, containingElement: UElement?): UMethod {
